+++ lib/leanengine.js
@@ -57,6 +57,14 @@ AV.Cloud.HttpsRedirect = function(options) {
   }
 }
 
+AV.Cloud.LeanCloudHeaders = function(options) {
+  if (options && options.framework == 'koa') {
+    return require('../middleware/leancloud-heaedrs-koa')(AV)(options);
+  } else {
+    return require('../middleware/leancloud-heaedrs')(AV)(options);
+  }
+}
+
 function createCloudFunctionRouter(options) {
   options = options || {};
 
@@ -67,7 +75,7 @@ function createCloudFunctionRouter(options) {
   cloudFunctions.use(bodyParser.json({limit: '20mb'}));
   cloudFunctions.use(bodyParser.text({limit: '20mb'}));
   cloudFunctions.use(require('../middleware/cors')());
-  cloudFunctions.use(require('../middleware/parse-leancloud-headers')(AV, {restrict: true}));
+  cloudFunctions.use(require('../middleware/leancloud-headers')(AV)({restrict: true}));
 
   cloudFunctions.use('/_ops/metadatas', function(req, res) {
     if (req.AV.authMasterKey) {
